auth module renamed.
