+++ front-end/OSGB/src/app/modules/auth/components/auth.component.scss
